+++ CHANGES.md
@@ -4,7 +4,7 @@ This changelog tracks features and fixes of image-attribution PHP library.
 
 ## Unreleased
 
-* new field `license_name`
+* new fields `license_name`, `name`, `mime`, `size`, `width`, `height`
 * extend example as better demo
 
 ## 0.2.0